JAR created with make for Lab 2 project Distributed Applications
Create an application in the Java programming language that takes URL as an argument, reads the 
resource on this URL, and stores it in a file. The name of file will be specified as the second argument.
Build the application and transfer the JAR to EC2 instance on AWS. Run the application and download 
the resources at the following URLs******
Submission 
You will submit the zipped project with the screenshot.