Update to write URL InputStream into a file.
I did not realize that the req. sheet wanted me to read into a file.. I have updated the code to construct an InputStream directly from opening a URL stream, then read the bytes into a buffer and write them to a file in a loop until the stream returns -1 (no input).